+++ mrr/src/test/kotlin/gov/cdc/dex/mrr/EventCodeClientTest.kt
@@ -1,0 +1,19 @@
+package gov.cdc.dex.mrr
+
+import org.junit.jupiter.api.Test
+
+
+
+internal class EventCodeClientTest {
+
+    @Test
+    fun loadEventMapsTest() {
+        val eventCodes  = EventCodeClient()
+        eventCodes.loadEventMaps()
+    }
+    @Test
+    fun loadGroupsTest() {
+        val groupClient  = EventCodeClient()
+        groupClient.loadGroups()
+    }
+}